Add scene test + fix + debug
Add a test for FastScene roundtrip.
-----------------------------------

You don't know if it works until you test it, and testing it revealed
that it was completely and utterly broken.

Fix a bug where the scene wasn't read correctly.
------------------------------------------------

The issue was in hierarchy.rs. In both `child` and `build`, we were
passing the same &mut[] to all children inside of the `map` closure.
As a result, only the first child was written to, which made it so only
one entity was properly saved.

This might have been the most complex bug I had to debug in rust

Add debug methods to Tables and Inlines
---------------------------------------

Due to the complexity of the bug, I had to find ways to print in a
meaningful ways the content of the FastScene. This motivated the
introduction of a few methods that prints internal state of Keys, Tables
and Inlines.